move rewards to top level package
