+++ pkg/core/server/rewards_internal_test.go
@@ -8,7 +8,7 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/AudiusProject/audiusd/pkg/core/rewards"
+	"github.com/AudiusProject/audiusd/pkg/rewards"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/labstack/echo/v4"
 	"github.com/stretchr/testify/require"
