+++ pkg/rewards/reward_attester_test.go
@@ -5,7 +5,7 @@ import (
 	"log"
 	"testing"
 
-	"github.com/AudiusProject/audiusd/pkg/core/rewards"
+	"github.com/AudiusProject/audiusd/pkg/rewards"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/stretchr/testify/require"
 )
